memguard: bugfix: idle notifier
idle notifier can be called in any intterupt or perhaps
cache maintenance traffic (it was true in the case of Cortex-A9)

Hence, starting a new period (excl5 mode) cause a problem as it may
be called many times in a short period of time.

This patch prevents any activities other than simply mask/unmask
the active_mask, which then later can be used by the overflow hander.